DISPLAY-1030: Fixed PlaylistSlide API
