+++ src/Serializer/RelationNormalizer.php
@@ -59,6 +59,7 @@ class RelationNormalizer implements NormalizerInterface, NormalizerAwareInterfac
         $data = ['@context' => '/contexts/'.$context['output']['name']];
 
         if (isset($context['operation_type']) && OperationType::SUBRESOURCE === $context['operation_type']) {
+            // FIXME
             $data['@id'] = $this->iriConverter->getSubresourceIriFromResourceClass($resourceClass, $context);
         } else {
             $path = $this->utils->getApiPlatformPathPrefix().$context['output']['name'].'s';
